+++ src/renderer/containers/TopNav/TopNavNotifications/TopNavNotifications.scss
@@ -1,21 +1,2 @@
 .TopNavNotifications {
 }
-
-.TopNavNotifications__menu {
-  background: var(--color-white);
-  border-radius: 3px;
-  box-shadow: 0 3px 12px 1px rgba(4, 34, 53, 0.24);
-  position: fixed;
-  right: 12px;
-  top: 40px;
-
-  h2 {
-    border-bottom: 1px solid var(--color-sail-gray-100);
-    padding: 12px 32px 18px;
-  }
-
-  &__payment {
-    border-bottom: 1px solid var(--color-sail-gray-100);
-    padding: 12px 12px 12px 32px;
-  }
-}
